seperating unit tests for trivy backend changes and then fixing black formatting
